[WIP] Adding Charka UI to old components
