+++ package.json
@@ -5,7 +5,11 @@
 	"main": "index.js",
 	"license": "MIT",
 	"dependencies": {
+		"@chakra-ui/react": "^2.4.9",
+		"@emotion/react": "^11.10.5",
+		"@emotion/styled": "^11.10.5",
 		"debounce": "^1.2.1",
+		"framer-motion": "^9.0.2",
 		"highcharts": "^9.3.2",
 		"highcharts-react-official": "^3.1.0",
 		"lint-staged": "^12.1.4",
